⭐ New Features 1. 新增 DictTypeEnum.kt 和 MenuTypeEnum.kt

+++ src/main/java/com/simonvonxcvii/turing/model/dto/MenuDTO.java
@@ -72,7 +72,7 @@ public class MenuDTO extends PageQuery {
      * 菜单排序
      */
     @NotNull(message = "菜单排序不能为空")
-    private Integer sort;
+    private Short sort;
     /**
      * 是否显示
      */
